Move refund wave helpers to their own module; dedupe test helper
Address CI and review feedback on the provider test hardening:

- jscpd flagged a duplicated console.error helper across the two provider
  test files. Fold it into setupErrorSpy as a shared `contains(needle)`
  method and use that from both suites.
- packByReferenceCount and combineRefundOutcomes were exported from
  provider.ts solely for tests, which AGENTS.md rejects as a test-only
  public surface. Extract them (with the RefundOutcome type) into a pure
  refunds/waves.ts module that provider.ts imports, giving them a real
  production consumer and following the repo's pure/IO separation.

Update the equivalent-mutant record to the helper's new home. Mutation
score over provider.ts + waves.ts stays 100% (66/66, 1 suppressed);
full precommit passes.
