+++ scripts/mutation/equivalent-mutants.txt
@@ -845,6 +845,6 @@ src/shared/logger.ts:314:27  = → +=   # errorPersistGuard.active is always fal
 src/shared/logger.ts:321:69  ?? → ||   # context.listingId is number|undefined and listing ids are positive autoincrement integers, so the LHS is never a falsy non-null value — ?? and || agree on every input
 src/shared/logger.ts:77:70  ?? → ||   # getRequestId's `getStore() ?? ""`: the RHS is "" and the only falsy string is "", so `x ?? ""` and `x || ""` agree on every string|undefined input
 
-# Admin refund provider (refunds/provider.ts) — one provably-equivalent survivor
-# from the packByReferenceCount run.
-src/features/admin/refunds/provider.ts:45:24  0 → 1   # packByReferenceCount's `currentCount` init: the first iteration always hits `!wave` (waves is empty, so waves[-1] is undefined), which short-circuits the `|| currentCount + refs > budget` before currentCount is read; every later iteration reassigns it first, so the initial value is never observed
+# Admin refund waves (refunds/waves.ts) — one provably-equivalent survivor from
+# the packByReferenceCount run.
+src/features/admin/refunds/waves.ts:16:24  0 → 1   # packByReferenceCount's `currentCount` init: the first iteration always hits `!wave` (waves is empty, so waves[-1] is undefined), which short-circuits the `|| currentCount + refs > budget` before currentCount is read; every later iteration reassigns it first, so the initial value is never observed
